QtSfml: Printing current position using QtControls
This required many changes in GameEngine.
Created GameState class, which contains all objects on screen.
This state allows to simply initialize and reset game (and save/load
it in the future). Moreover this allows to easily read current state
and print it in QLabel.
To synchronize printing with animation, timer had to be moved
from GameEngine to MainWindow.