Core: Variable range random number generator
Maze generation requires to pick random tile from unvisited adjacent
tiles. Currently used solution with srand() was not the best.
The number of rooms varies between zero and three so a variable
range random number generator was needed.

It may be an overkill to create separate class for this task but
I wanted to write test for this to be sure that it works correctly