QSD: Maze - Generator interface
Preparation for multiple maze generation algorithms in future.
An interface is required to create multiple algorithms and change
them in runtime.
Generation algorithms were moved to Core directory, because they
are not dependent on any Qt or SFML class and I may use them in
future in other projects (maze solving algorithms)