Jrnl: Split logic between model and controller
Model will be responsible only for holding the data, page controller
will be responsible for interacting with user. The goas is to do
something like this:
User presses button -> controller receives signal -> model is modified

I Added some changes in database in this commit too. It got too
complicated to split it now.